Added CompilerInputPath to nodes so they can compute base dir for single files.
Added CompilerInputPath to the list of values that get passed to node
creation so that in GatherDynamicDependencies, the node can compute a
baseDir for file nodes.  This will allow better locations for obj files
in some cases.  This is useful for build config generators, which can
more easily supply a list of files instead of dirs and filters.  Note
that the generator should also supply a bogus CompilerInputPattern such
as 'ignoreall' if the intention is to use the InputPath to base the
explicit list of files.